Add database CRUD operations and update model
Implemented CRUD functions for RelatorioSocial in database.py and updated the engajamento field in models.py from Integer to Float for more precise data storage. Also added a placeholder for social_api.py service.

+++ backend/app/models.py
@@ -21,5 +21,5 @@ class RelatorioSocial(Base):
     seguidores_fim = Column(Integer, nullable=False)
     publicacoes = Column(Integer, nullable=False)
     alcance_total = Column(Integer, nullable=False)
-    engajamento = Column(Integer, nullable=False)
+    engajamento = Column(Float, nullable=False)
     origem = Column(Enum(OrigemDadosEnum), nullable=False)
